Changed the cutoff in the x-direction (long direction)
Now the beam stops at the value of the Gaussian radius in the
x-direction. This is done so as to simulate the effect that if an atom
is flying off-axis, then it has to be slowed by the other beams, but if
it's outside of the range of those beams, it will escape. This is
simulated by setting the intensity to 0 past Gaussian radius, this
essentially means that the beams cannot slow an atom with such velocity
flying in their direction before they escape the range of the
perpendicular beams