Wrote down the code with experimental params
blueMOT2D_experimental.py should describe well the experimental
situation that we have. I reduced the number of varied initial
conditions to make the code run faster and removed all the stuff for
saving and labeling in HDF5 files. Now it’s directly for calculating
and plotting, therefore the calculations must be reasonably fast